moved pg.py stuff into dbms/postgres/cli.py
